moved files around, changed error codes and removed some logging
