+++ src/main/java/com/kijiri/aurora/api/controller/request/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.kijiri.aurora.api.request;
+package com.kijiri.aurora.api.controller.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
