+++ src/main/java/com/kijiri/aurora/api/controller/request/RegistrationRequest.java
@@ -1,4 +1,4 @@
-package com.kijiri.aurora.api.request;
+package com.kijiri.aurora.api.controller.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
